Use a WHERE clause instead of the EXCEPTION
The EXCEPTION as implemented applied to the 'batch' insert, meaning
if it was hit, we would swallow the error but the entire insert would
fail -- so the remaining, perfectly valid data would not get loaded.

Working in a LOOP and applying EXCEPTION to the individual inserts
would be an enormous performance drag, so this commit instead uses
a much simplere WHERE clause to ensure that a row is not inserted if
its url matches an existing row with a different foreign id.