Use DISTINCT ON to remove duplicates in a single batch
The previous commit works to prevent inserting a new record that matches an
existing direct_url in the Catalog, but it fails to catch the case where multiple
records with the same direct_url are returned by the API in the same batch.

This commit adds a DISTINCT ON clause to remove these duplicates first.