Make the ExternalDAGsSensor deferrable
We need to run the data refresh DAGs in a pool with only one slot
in order to make sure no trigger tasks run simultaneously. This
means that the sensor tasks MUST free up the worker slot while
waiting, or the entire system will simply deadlock on a DAG that
is waiting for its siblings to finish.