Configure poke interval
I don't think it's really necessary to have different poke intervals
per media type for the wait task, and it's a slightly confusing
attribute in the dataclass since it's specific to a particular task.
Now it's configurable for the Sensor, but the dag factory defines a
single poke_interval for all data refresh DAGs.