Don't refresh metrics is option is explicitly configured to False
If `force_refresh_metrics` is explictly configured to False (rather
than omitted), then do not refresh the popularity metrics even if
this is the first successful run of the month.

This option could be helpful if, for example, the first dagrun of the
month succeeds during the popularity steps but fails during the data
refresh. When we manually re-run the DAG, we can save time by skipping
this step.