Remove clean step from load_local_data_to_intermediate_table
As far as I can tell this method is tested but never used anywhere.
This should mirror `load_s3_data_to_intermediate_table` and only
handle the loading steps, not the cleaning steps.

If this method *is* used elsewhere, it will need to be updated to
call the cleaning steps separately.